arkhe-os-genesis-v1.0/src/glp_interface/requirements.txt
torch==2.2.0
flask==3.0.0
numpy==1.24.0
scikit-learn==1.3.0
flask-socketio==5.3.6
pyserial==3.5
fpdf==1.7.2
python-dotenv==1.0.0
python-osc==1.8.3
openai==1.12.0
